Move `Outline` to `sbr-util`
