+++ sbr-util/src/math/outline.rs
@@ -1,9 +1,7 @@
-#![expect(dead_code)] // tessellation
-
 use std::fmt::Debug;
 
-use text_sys::{FT_Outline, FT_Vector, FT_CURVE_TAG_CONIC, FT_CURVE_TAG_CUBIC, FT_CURVE_TAG_ON};
-use util::{fmt_from_fn, math::*};
+use super::*;
+use crate::fmt_from_fn;
 
 pub struct OutlineBuilder {
     outline: Outline,
@@ -124,87 +122,6 @@ impl Outline {
             points: Vec::new(),
             segments: Vec::new(),
         }
-    }
-
-    pub unsafe fn from_freetype(ft: &FT_Outline) -> Self {
-        let mut first = 0;
-        let mut builder = OutlineBuilder::new();
-        let contours = std::slice::from_raw_parts(ft.contours, ft.n_contours as usize);
-        let points = std::slice::from_raw_parts(ft.points, ft.n_points as usize);
-        let tags = std::slice::from_raw_parts(ft.tags, ft.n_points as usize);
-
-        // TODO: Convert FT_CURVE_TAG* to u8 in text-sys
-        for last in contours.iter().map(|&x| x as usize) {
-            // FT_Pos in FT_Outline seems to be 26.6
-            let to_point = |vec: FT_Vector| {
-                Point2f::new(
-                    vec.x as f32 * 2.0f32.powi(-6),
-                    // FreeType uses a "Y axis at the bottom" coordinate system,
-                    // flip that to match ours
-                    -vec.y as f32 * 2.0f32.powi(-6),
-                )
-            };
-
-            let midpoint =
-                |a: Point2f, b: Point2f| Point2f::new((a.x + b.x) / 2.0, (a.y + b.y) / 2.0);
-
-            let mut last_tag;
-            let mut final_degree = SegmentDegree::Linear;
-            let mut add_range = first..last + 1;
-            if (tags[first] & 0b11) != FT_CURVE_TAG_ON as u8 {
-                if (tags[last] & 0b11) == FT_CURVE_TAG_CONIC as u8 {
-                    builder.add_point(midpoint(to_point(points[first]), to_point(points[last])));
-                    last_tag = FT_CURVE_TAG_ON as u8;
-                    final_degree = SegmentDegree::Quadratic;
-                } else {
-                    assert_eq!(tags[last] & 0b11, FT_CURVE_TAG_ON as u8);
-                    builder.add_point(to_point(points[last]));
-                    last_tag = tags[last] & 0b11;
-                    add_range.end -= 1;
-                }
-            } else {
-                builder.add_point(to_point(points[first]));
-                last_tag = tags[first] & 0b11;
-                add_range.start += 1;
-                if tags[last] & 0b11 == FT_CURVE_TAG_CUBIC as u8 {
-                    final_degree = SegmentDegree::Cubic;
-                } else if tags[last] & 0b11 == FT_CURVE_TAG_CONIC as u8 {
-                    final_degree = SegmentDegree::Quadratic;
-                }
-            }
-
-            for (&point, &tag) in points[add_range.clone()].iter().zip(tags[add_range].iter()) {
-                let tag = tag & 0b11;
-                let point = to_point(point);
-
-                if tag == FT_CURVE_TAG_ON as u8 {
-                    if last_tag == FT_CURVE_TAG_ON as u8 {
-                        builder.add_segment(SegmentDegree::Linear);
-                    } else if last_tag == FT_CURVE_TAG_CONIC as u8 {
-                        builder.add_segment(SegmentDegree::Quadratic);
-                    } else {
-                        builder.add_segment(SegmentDegree::Cubic);
-                    }
-                }
-
-                if tag == FT_CURVE_TAG_CONIC as u8 && last_tag == FT_CURVE_TAG_CONIC as u8 {
-                    let last = *builder.points().last().unwrap();
-                    builder.add_point(midpoint(last, point));
-                    builder.add_segment(SegmentDegree::Quadratic);
-                }
-
-                last_tag = tag;
-                builder.add_point(point);
-            }
-
-            builder.add_segment(final_degree);
-            builder.close_contour();
-            first = last + 1;
-        }
-
-        assert_eq!(first, points.len());
-
-        builder.build()
     }
 
     #[inline(always)]
